feat: description, subtask hooks 적용
