+++ apps/client/src/features/task/subtask/types.ts
@@ -4,7 +4,34 @@ export type Subtask = {
   completed: boolean;
 };
 
+// create
+export interface CreateSubtaskResult {
+  subtask: {
+    id: number;
+    content: string;
+    completed: boolean;
+  };
+}
+
+// update
 export interface UpdateSubtaskDto {
   content?: string;
   completed?: boolean;
+}
+
+export interface UpdateSubtaskResult {
+  subtask: {
+    id: number;
+    content: string;
+    completed: boolean;
+  };
+}
+
+// delete
+export interface DeleteSubtaskResult {
+  subtask: {
+    id: number;
+    content: string;
+    completed: boolean;
+  };
 }
